backend/requirements.txt
asgiref==3.7.2
astroid==3.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.4
defusedxml==0.8.0rc2
dill==0.3.8
Django==3.2.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.6
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
pillow==10.2.0
platformdirs==4.2.0
pycodestyle==2.11.1
pycparser==2.21
PyJWT==2.8.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tomlkit==0.12.3
urllib3==2.2.1
webcolors==1.13
